+++ src/main/java/domain/deck/CardDeck.java
@@ -1,4 +1,4 @@
-package domain;
+package domain.deck;
 
 import java.util.Queue;
