+++ src/main/java/domain/participants/Dealer.java
@@ -1,4 +1,4 @@
-package domain;
+package domain.participants;
 
 
 public class Dealer extends Player {
