+++ src/main/java/strategy/ShuffleCardsStrategy.java
@@ -1,6 +1,6 @@
 package strategy;
 
-import domain.Card;
+import domain.deck.Card;
 
 import java.util.Collections;
 import java.util.List;
