+++ src/test/java/domain/NameTest.java
@@ -2,12 +2,11 @@ package domain;
 
 import static org.assertj.core.api.Assertions.*;
 
+import domain.participants.Name;
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
 import org.junit.jupiter.params.ParameterizedTest;
 import org.junit.jupiter.params.provider.NullAndEmptySource;
-
-import static org.junit.jupiter.api.Assertions.*;
 
 public class NameTest {
     @Test
